+++ src/opm/parser/eclipse/EclipseState/Schedule/TimeMap.cpp
@@ -331,11 +331,11 @@ namespace Opm {
     }
 
 
-    std::time_t TimeMap::forward(std::time_t t, long seconds) {
+    std::time_t TimeMap::forward(std::time_t t, int64_t seconds) {
         return t + seconds;
     }
 
-    std::time_t TimeMap::forward(std::time_t t, int hours, int minutes, long seconds) {
+    std::time_t TimeMap::forward(std::time_t t, int64_t hours, int64_t minutes, int64_t seconds) {
         return t + seconds + minutes * 60 + hours * 3600;
     }
 }
